Revert "update: 웨크웨어 데이터 작업중"
This reverts commit 8f4496c04fc8ef28ca80bee532e1972a38fff5eb.
