+++ src/main/java/board/workwear/dao/WorkwearDAO.java
@@ -6,6 +6,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+import board.workwear.dto.WorkwearDTO;
 import common.DBConnectionUtil;
 
 public class WorkwearDAO {
